Add framework for study render
Listens to changes api for procedure steps that it is able
to process and if found downloads the corresponding studies.
Next step is to render the studies and then to save out
png or jpg images of the rendering and convert them to
secondary captures for the study.  These can then be pushed
back to chronicle.